requirements.txt
attrs==19.3.0
Automat==0.8.0
cachetools==3.0.0
certifi==2018.11.29
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
google-api-core==1.7.0
google-auth==1.6.2
google-auth-oauthlib==0.2.0
google-cloud-bigquery==1.8.1
google-cloud-core==0.29.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.2
numpy==1.16.0
oauthlib==3.0.1
pandas==0.23.4
pandas-gbq==0.9.0
parsel==1.5.2
Protego==0.1.16
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pydata-google-auth==0.1.1
PyDispatcher==2.0.5
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
python-dateutil==2.7.5
pytz==2018.9
queuelib==1.5.0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.12.0
Twisted==19.10.0
urllib3==1.24.1
w3lib==1.21.0
zope.interface==4.7.1
